Remove unused RestClient move PetFinderHttpClient into network package
