+++ app/src/main/java/com/codepath/apps/critterfinder/CritterFinderApplication.java
@@ -7,10 +7,8 @@ import android.content.Context;
  * including the image cache in memory and on disk. This also adds a singleton
  * for accessing the relevant rest client.
  *
- *     RestClient client = CritterFinderApplication.getRestClient();
- *     // use client to send requests to API
- *
  */
+
 public class CritterFinderApplication extends com.activeandroid.app.Application {
 	private static Context context;
 
@@ -20,7 +18,4 @@ public class CritterFinderApplication extends com.activeandroid.app.Application 
 		CritterFinderApplication.context = this;
 	}
 
-	public static RestClient getRestClient() {
-		return (RestClient) RestClient.getInstance(RestClient.class, CritterFinderApplication.context);
-	}
 }
