+++ app/src/main/java/com/codepath/apps/critterfinder/network/PetFinderHttpClient.java
@@ -1,7 +1,6 @@
-package com.codepath.apps.critterfinder;
+package com.codepath.apps.critterfinder.network;
 
 import com.loopj.android.http.AsyncHttpClient;
-import com.loopj.android.http.AsyncHttpResponseHandler;
 import com.loopj.android.http.JsonHttpResponseHandler;
 import com.loopj.android.http.RequestParams;
